Move entrypoint interface types to point of consumpion
Move entrypoint interface types to point of consumpion
