+++ internal/pkg/impl/entrypoint/app/workflow_runner.go
@@ -1,4 +1,4 @@
-package workflow
+package app
 
 import (
 	"io"
